% Please edit documentation in R/16_DPprior_fit.R
\name{confidence_to_vif_fit}
\alias{confidence_to_vif_fit}
\title{Convert a qualitative confidence level to a variance inflation factor}
\usage{
confidence_to_vif_fit(confidence)
}
\arguments{
\item{confidence}{One of low, medium, or high.}
}
\value{
The corresponding variance inflation factor.
}
\description{
Convert a qualitative confidence level to a variance inflation factor
}
\keyword{internal}
