% Please edit documentation in R/09_weights_rho.R
\name{cv_rho}
\alias{cv_rho}
\title{Coefficient of Variation of rho}
\usage{
cv_rho(a, b, M = .QUAD_NODES_DEFAULT)
}
\arguments{
\item{a}{Numeric; shape parameter of the Gamma prior on alpha (a > 0).}

\item{b}{Numeric; rate parameter of the Gamma prior on alpha (b > 0).}

\item{M}{Integer; number of quadrature nodes. Default is 80.}
}
\value{
Numeric; coefficient of variation.
}
\description{
Computes CV(rho) = SD(rho) / E(rho) under alpha ~ Gamma(a, b).
}
\examples{
cv_rho(a = 2, b = 1)
cv_rho(a = 1.6, b = 1.22)

}
\seealso{
\code{\link{mean_rho}}, \code{\link{var_rho}}

Other co_clustering:
\code{\link[=mean_rho]{mean_rho()}},
\code{\link[=mean_rho_given_alpha]{mean_rho_given_alpha()}},
\code{\link[=var_rho]{var_rho()}},
\code{\link[=var_rho_given_alpha]{var_rho_given_alpha()}}
}
\concept{co_clustering}
