% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_moments_conditional.R
\docType{data}
\name{.ALPHA_SMALL_MOMENTS}
\alias{.ALPHA_SMALL_MOMENTS}
\title{Threshold for Small Alpha Guard}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.ALPHA_SMALL_MOMENTS
}
\description{
Below this value, we use limiting behavior directly to avoid
numerical cancellation in digamma/trigamma expressions.
}
\keyword{internal}
