% Please edit documentation in R/18_visualization.R
\name{.dpprior_dual_comparison_table}
\alias{.dpprior_dual_comparison_table}
\title{Dual comparison summary table}
\usage{
.dpprior_dual_comparison_table(dual, summ_K, summ_dual, base_size = 11)
}
\description{
Dual comparison summary table
}
\keyword{internal}
