% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_visualization.R
\name{.dpprior_get_K_pmf}
\alias{.dpprior_get_K_pmf}
\title{Hierarchical K PMF Computation with Fallbacks}
\usage{
.dpprior_get_K_pmf(J, a, b, M = NULL)
}
\description{
Hierarchical K PMF Computation with Fallbacks
}
\keyword{internal}
