% Please edit documentation in R/00_constants.R
\name{.MAX_J_DEFAULT}
\alias{.MAX_J_DEFAULT}
\title{Maximum Supported Sample Size}
\usage{
.MAX_J_DEFAULT
}
\description{
Default maximum value of J (sample size) supported by the package.
Pre-computed Stirling number tables are limited to this size.
}
\keyword{internal}
