% Please edit documentation in R/00_constants.R
\name{.QUAD_NODES_DEFAULT}
\alias{.QUAD_NODES_DEFAULT}
\title{Default Number of Quadrature Nodes}
\usage{
.QUAD_NODES_DEFAULT
}
\description{
Default number of Gauss-Laguerre quadrature nodes for
numerical integration against Gamma distributions.
}
\keyword{internal}
