% Please edit documentation in R/07_jacobian.R
\name{.QUAD_NODES_VERIFICATION}
\alias{.QUAD_NODES_VERIFICATION}
\title{Default Quadrature Nodes for Verification}
\usage{
.QUAD_NODES_VERIFICATION
}
\description{
Higher M value recommended for Jacobian verification due to
slower convergence of score-weighted integrands.
}
\keyword{internal}
