% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_constants.R
\docType{data}
\name{.TOL_NEWTON}
\alias{.TOL_NEWTON}
\title{Newton Method Convergence Tolerance}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.TOL_NEWTON
}
\description{
Convergence tolerance for Newton's method in moment matching.
}
\keyword{internal}
