% Please edit documentation in R/00_constants.R
\name{.TOL_PROJECTION_BUFFER}
\alias{.TOL_PROJECTION_BUFFER}
\title{Feasibility Projection Buffer}
\usage{
.TOL_PROJECTION_BUFFER
}
\description{
Small buffer for projecting to the feasible region of the
negative binomial approximation.
}
\keyword{internal}
