% Please edit documentation in R/18_visualization.R
\name{DPprior_colors}
\alias{DPprior_colors}
\title{DPprior Color Palette}
\usage{
DPprior_colors()
}
\value{
A named list with color values.
}
\description{
DPprior Color Palette
}
\seealso{
Other visualization:
\code{\link[=plot_K_prior]{plot_K_prior()}},
\code{\link[=plot_alpha_prior]{plot_alpha_prior()}},
\code{\link[=plot_dual_comparison]{plot_dual_comparison()}},
\code{\link[=plot_dual_dashboard]{plot_dual_dashboard()}},
\code{\link[=plot_prior_dashboard]{plot_prior_dashboard()}},
\code{\link[=plot_tradeoff_curve]{plot_tradeoff_curve()}},
\code{\link[=plot_tradeoff_dashboard]{plot_tradeoff_dashboard()}},
\code{\link[=plot_w1_prior]{plot_w1_prior()}},
\code{\link[=theme_DPprior]{theme_DPprior()}}
}
\concept{visualization}
