% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/17_s3_methods.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check}

\item{y}{Default value if x is NULL}
}
\value{
x if not NULL, otherwise y
}
\description{
Null-Coalescing Operator
}
\keyword{internal}
