% Please edit documentation in R/06_pmf_marginal.R
\name{mean_K_from_marginal_pmf}
\alias{mean_K_from_marginal_pmf}
\title{Mean of Marginal K from PMF}
\usage{
mean_K_from_marginal_pmf(J, a, b, logS, M = .QUAD_NODES_DEFAULT)
}
\arguments{
\item{J}{Integer; sample size.}

\item{a}{Numeric; shape parameter of Gamma prior.}

\item{b}{Numeric; rate parameter of Gamma prior.}

\item{logS}{Matrix; pre-computed log-Stirling matrix.}

\item{M}{Integer; number of quadrature nodes.}
}
\value{
Numeric; marginal mean.
}
\description{
Computes \eqn{E[K_J \mid a, b]} from the marginal PMF.
}
\examples{
\dontrun{
logS <- compute_log_stirling(50)
mean_K_from_marginal_pmf(50, 1.5, 0.5, logS)

}
}
\keyword{internal}
