% Please edit documentation in R/14_diagnostics.R
\name{print.DPprior_diagnostics}
\alias{print.DPprior_diagnostics}
\title{Print Method for DPprior_diagnostics Objects}
\usage{
\method{print}{DPprior_diagnostics}(x, ...)
}
\arguments{
\item{x}{A \code{DPprior_diagnostics} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for DPprior_diagnostics Objects
}
