% Please edit documentation in R/13_error_bounds.R
\name{print.DPprior_error_bounds}
\alias{print.DPprior_error_bounds}
\title{Print Method for DPprior_error_bounds Objects}
\usage{
\method{print}{DPprior_error_bounds}(x, ...)
}
\arguments{
\item{x}{An object of class "DPprior_error_bounds".}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Displays a formatted summary of the A1 approximation error analysis.
}
