% Please edit documentation in R/08_weights_w1.R
\name{print.w1_summary}
\alias{print.w1_summary}
\title{Print Method for w1_summary Objects}
\usage{
\method{print}{w1_summary}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "w1_summary".}

\item{digits}{Integer; number of digits for printing.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for w1_summary Objects
}
