% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/18_visualization.R
\name{.dpprior_compute_summary}
\alias{.dpprior_compute_summary}
\title{Compute Summary Statistics}
\usage{
.dpprior_compute_summary(
  fit = NULL,
  a = NULL,
  b = NULL,
  J = NULL,
  ci_level = 0.95
)
}
\description{
Compute Summary Statistics
}
\keyword{internal}
