% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_constants.R
\docType{data}
\name{.MAX_J_DEFAULT}
\alias{.MAX_J_DEFAULT}
\title{Maximum Supported Sample Size}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.MAX_J_DEFAULT
}
\description{
Default maximum value of J (sample size) supported by the package.
Pre-computed Stirling number tables are limited to this size.
}
\keyword{internal}
