% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_constants.R
\docType{data}
\name{.QUAD_NODES_DEFAULT}
\alias{.QUAD_NODES_DEFAULT}
\title{Default Number of Quadrature Nodes}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.QUAD_NODES_DEFAULT
}
\description{
Default number of Gauss-Laguerre quadrature nodes for
numerical integration against Gamma distributions.
}
\keyword{internal}
