% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_jacobian.R
\docType{data}
\name{.QUAD_NODES_VERIFICATION}
\alias{.QUAD_NODES_VERIFICATION}
\title{Default Quadrature Nodes for Verification}
\format{
An object of class \code{integer} of length 1.
}
\usage{
.QUAD_NODES_VERIFICATION
}
\description{
Higher M value recommended for Jacobian verification due to
slower convergence of score-weighted integrands.
}
\keyword{internal}
