% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_constants.R
\docType{data}
\name{.TOL_PROJECTION_BUFFER}
\alias{.TOL_PROJECTION_BUFFER}
\title{Feasibility Projection Buffer}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
.TOL_PROJECTION_BUFFER
}
\description{
Small buffer for projecting to the feasible region of the
negative binomial approximation.
}
\keyword{internal}
