% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPprior-package.R
\docType{package}
\name{DPprior-package}
\alias{DPprior}
\alias{DPprior-package}
\title{DPprior: Dirichlet Process Prior Elicitation}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides calibration tools for the concentration parameter of Dirichlet process priors, enabling principled prior specification through cluster count and weight distribution targets. Implements closed-form approximations and exact numerical methods for mapping user beliefs about expected cluster counts to Gamma hyperparameters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JoonHoLee/DPprior}
  \item Report bugs at \url{https://github.com/JoonHoLee/DPprior/issues}
}

}
\author{
\strong{Maintainer}: JoonHo Lee \email{jlee296@ua.edu} (\href{https://orcid.org/0009-0006-4019-8703}{ORCID})

}
\keyword{internal}
