% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_moments_marginal.R
\name{K_moments}
\alias{K_moments}
\title{Convenience Wrapper for Marginal Moments}
\usage{
K_moments(J, a, b, M = .QUAD_NODES_DEFAULT)
}
\arguments{
\item{J}{Integer; sample size (positive integer >= 1).}

\item{a}{Numeric; shape parameter of Gamma prior (> 0).}

\item{b}{Numeric; rate parameter of Gamma prior (> 0).}

\item{M}{Integer; number of quadrature nodes (default: 80).}
}
\value{
Named numeric vector \code{c(mean = ..., var = ...)}.
}
\description{
Returns marginal mean and variance as a named numeric vector.
}
\examples{
K_moments(50, 2.0, 1.0)

}
\seealso{
\code{\link{exact_K_moments}} for full output

Other marginal_K: 
\code{\link{cdf_K_marginal}()},
\code{\link{exact_K_moments}()},
\code{\link{mode_K_marginal}()},
\code{\link{pmf_K_marginal}()},
\code{\link{quantile_K_marginal}()},
\code{\link{summary_K_marginal}()}
}
\concept{marginal_K}
