% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14_diagnostics.R
\name{print.DPprior_diagnostics}
\alias{print.DPprior_diagnostics}
\title{Print Method for DPprior_diagnostics Objects}
\usage{
\method{print}{DPprior_diagnostics}(x, ...)
}
\arguments{
\item{x}{An object of class "DPprior_diagnostics".}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for DPprior_diagnostics Objects
}
