% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_weights_rho.R
\name{print.rho_summary}
\alias{print.rho_summary}
\title{Print Method for rho_summary Objects}
\usage{
\method{print}{rho_summary}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "rho_summary".}

\item{digits}{Integer; number of digits for printing.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for rho_summary Objects
}
