% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_weights_rho.R
\name{rrho}
\alias{rrho}
\title{Random Generation from rho Distribution}
\usage{
rrho(n, a, b, n_sticks = 500L)
}
\arguments{
\item{n}{Integer; number of samples to generate.}

\item{a}{Numeric; shape parameter of the Gamma prior on alpha (a > 0).}

\item{b}{Numeric; rate parameter of the Gamma prior on alpha (b > 0).}

\item{n_sticks}{Integer; number of sticks for truncation. Default is 500.}
}
\value{
Numeric vector of length n; random samples from the rho distribution.
}
\description{
Generates random samples from the rho = sum w_h^2 distribution by
stick-breaking simulation.
}
\details{
Uses the hierarchical representation:
\enumerate{
\item alpha ~ Gamma(a, b)
\item v_h | alpha ~ Beta(1, alpha) independently for h = 1, ..., n_sticks
\item w_1 = v_1, w_h = v_h * prod(1 - v_l) for l < h
\item rho = sum w_h^2
}

Useful for Monte Carlo validation of analytical formulas.
}
\examples{
set.seed(42)
rho_samples <- rrho(1000, a = 2, b = 1)
mean(rho_samples)
mean_rho(a = 2, b = 1)

}
\seealso{
\code{\link{rw1}} for w1 random generation
}
