% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_moments_conditional.R
\name{summary_K_given_alpha}
\alias{summary_K_given_alpha}
\title{Summary of Conditional Moments}
\usage{
summary_K_given_alpha(J, alpha)
}
\arguments{
\item{J}{Sample size (integer >= 1).}

\item{alpha}{Concentration parameter (scalar).}
}
\value{
A list with components: J, alpha, mean, var, sd, cv, dispersion, dmean_dalpha.
}
\description{
Returns a comprehensive summary of the conditional distribution of K given alpha.
}
\examples{
\dontrun{
summary_K_given_alpha(50, 2.0)

}
}
\keyword{internal}
