% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_weights_w1.R
\name{var_w1}
\alias{var_w1}
\title{Variance of w₁}
\usage{
var_w1(a, b, M = .QUAD_NODES_DEFAULT)
}
\arguments{
\item{a}{Numeric; shape parameter of the Gamma prior on α (a > 0).}

\item{b}{Numeric; rate parameter of the Gamma prior on α (b > 0).}

\item{M}{Integer; number of quadrature nodes. Default is 80.}
}
\value{
Numeric; Var(w₁).
}
\description{
Computes Var(w₁ | a, b) using the law of total variance.
}
\details{
Uses the law of total variance:
\deqn{Var(w_1) = E[Var(w_1 | \alpha)] + Var(E[w_1 | \alpha])}

where w₁ | α ~ Beta(1, α), so:
\itemize{
\item E(w₁ | α) = 1/(1+α)
\item Var(w₁ | α) = α / ((1+α)²(2+α))
}
}
\examples{
var_w1(a = 2, b = 1)       # ~0.090
var_w1(a = 1.6, b = 1.22)  # ~0.105

}
\references{
Lee, J. (2025). RN-06: Dual-Anchor Design II.
}
\seealso{
\code{\link{mean_w1}}, \code{\link{summary_w1}}
}
